+++ Changelog.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - auto_forget option: automatically mark added card as unknown
 - open a card in browser by clicking "View added note (alt-v)" after adding it in yomichan
+### Fixed
+- improve error message when duplicate detection is on
 
 
 ## [0.4.1] - 2022-08-21